requirements.txt
fastapi
uvicorn
joblib
pandas
numpy
scikit-learn
tensorflow
keras
Pillow
gunicorn
google-cloud-storage
requests
python-multipart
jinja2